don't remove spaces from exact strings
The original reason for doing this was because the chat reader can't
read spaces, so chat messages don't have spaces in them and the match
text needs to reflect that. But doing this made popup alerts basically
unusable, so whatever. In future hopefully I will make the chat module
support reading spaces.